Remove model and use SQLite cache directly everywhere
The previous strategy for caching IMAP data from the server had a couple
of problems. Firstly, the system was supposed to consist of 3 layers:

1. In-memory cache layer (the model)
2. On-disk persistence layer (SQLite database)
3. Network communication

However, when you consider the rest of the program, the reality is much
messier:

1. Terminal output through curses
2. Widget sorteddicts which duplicate all of the keys in the model
3. The model
4. SQLite's internal caching
5. SQLite
6. Network communication

Secondly, populating the model involved unscalable scans of the SQLite
database. In particular, opening a mailbox would require reading every
row in the gmail_mailbox_uids table corresponding to a given mailbox and
then the row in the gmail_messages corresponding to each of those rows.
This is unacceptable for large mailboxes.

So, rather than trying to implement some sort of background database
scanning, this commit eliminates the model entirely. The widgets for the
mailbox sidebar, index view, and message view now access the database
directly. The updated "reality" picture is:

1. curses
2. SQLite's cache
3. SQLite
4. Network

The database schema is essentially the same as it was before with added
indices to make the sorting in the widgets possible. SQLite triggers
also replace the model callbacks.

The operation unit tests will have to be rewritten to accommodate this
refactoring in a future commit.