Fix message deleting during fetching logic
The idea with the call to Cache.delete_fetching_missing() in
GmailFetchMessagesOperation._handle_tagged_fetch_uids() is that at that
point, all of the UIDs that still exist in the range that we're fetching
should already be in the temporary fetching table. However, the seq_set
is not necessarily a single continuous range. If it isn't, the call to
delete_fetching_missing() would delete all of the messages in between
any of the ranges. The fix is that the last_uid passed to the
GmailFetchMessagesOperation should be the lowest UID that we have EVER
fetched, not the UID of the sequence number one above the highest
sequence number we're about to fetch.