Use ESEARCH to populate UIDs instead of batching
This is a major change to the syncing algorithm. By paying the upfront
cost of getting a list of all UIDs, we can avoid one round trip per
batch of messages to fetch the UIDs. Additionally, we can also use UID
FETCH now, which saves us another round trip, the NOOP in between
batches.